refactor(settings): the per-type schema editor becomes a component two hosts can open
Slice 2a of the ER diagram. No behaviour change; this is the move that lets the
Brain Overview open the SAME editor in place instead of sending an operator to
Space Settings for a one-field change and back.

The coupling it removes
------------------------

The editor body was ~224 template lines inside `space-schema-tab.component.ts`,
bound to `SpaceSettingsState` through `state.typeState(kt, name)` on roughly
thirty expressions. `SpaceSettingsState` is `@Injectable()` with no
`providedIn`, so it is provided by the settings page and the Brain page cannot
inject it; root-providing it would turn per-space editing state into a
cross-page singleton, which is worse than the problem.

The component takes a DRAFT and edits it in place. Its edits delegate to the
pure operations extracted in the previous slice, so the two hosts cannot
diverge.

The boundary, and why it is there
----------------------------------

Only the schema-EDITING body moved: naming pattern, retention, and the property
table with its constraint rows and enum chips.

Export, save-to-library, unlink and remove STAY on the tab's header. Those are
about the schema library and about deleting a type, not about editing one, and
the Overview panel has no business offering them — a dialog that could delete a
type would be answerable for something no caller asked it to do. `unlink` is
emitted rather than performed, because the read-only notice that offers it lives
in the body while the action belongs to the host.

It also does not save, and it does not read the space config. The two hosts
persist differently and must keep doing so — settings is staged, the panel is
immediate — and the inherited retention window arrives as an input.

What it DOES own is which property rows are expanded. That is view state, not
schema: the tab kept one set spanning every type it had open, and a dialog
editing one type has no use for that. It is also why `addProp` returns the key
it added rather than a boolean.

Styles moved to `schema-styles.ts` and are imported by BOTH. The editor takes
the whole set rather than a hand-split subset: a few list-side rules go unused
there, which costs nothing, where splitting by hand risks leaving a rule behind
for a control that still uses it. Two copies would drift the moment either side
was touched, and the drift would be invisible until somebody compared the
settings page against the Overview dialog side by side.

Three things the tooling caught, not me
----------------------------------------

  - `propConstraintSummary` was still on the tab. The AOT compiler named it in
    seconds; it moved with the rows it serves.
  - `HscrollTopDirective` became unused on the tab once the table left. NG8113
    said so; removed.
  - `type-schema-editable.test.js` failed, and it was RIGHT to: it asserts a
    human can actually set `retentionDays`, and it names the file holding the
    binding. Re-pointed at the component rather than relaxed — that gate exists
    because `retention` once shipped end to end with no control at all, and a
    canary operator found it by looking for the input the docs promised.

That last one nearly went wrong in a way worth recording. My first re-point was
written through a shell that ate the template literals, leaving
`assert.match(body, new RegExp(), )` — which passes against anything. It went
green and would have shipped a gate checking nothing, on the exact test whose
own header warns about a floor that keeps an empty parse from passing. Rewritten
from a file, and mutation-checked: renaming the binding fails it by name.

977 client tests green, preflight green.
