+++ client/src/app/pages/settings/space-schema-tab.component.ts
@@ -25,101 +25,16 @@ import { SpaceSettingsState, emptyTypeSchemaState, typeSchemaFromState, type Typ
 import { SchemaApi } from '../../core/schema-api.service';
 import { ToastService } from '../../core/toast.service';
 import { KnowledgeType, PropertySchema, SchemaLibraryEntry, TypeSchema, recordTtlWindows } from '../../core/api.types';
-import { HscrollTopDirective } from '../../shared/hscroll-top.directive';
 import { ErrorStateComponent } from '../../shared/error-state.component';
 import { httpErrorReason } from '../../core/http-error';
 
-const SCHEMA_MD_STYLES = `
-/* A floor, so this row cannot collapse and drag the master/detail grid up with it. The row's height is
-   otherwise stable by construction now: one hint string for all four collections, differing by a single
-   field name, so it wraps the same way whichever tab is open. That is what stops the add control below
-   from moving when you switch category. */
-.sch-head-row { display:flex; align-items:center; justify-content:space-between; gap:12px; flex-wrap:wrap;
-  min-height:20px; }
-.val-controls { display:inline-flex; align-items:center; gap:16px; flex-wrap:wrap; }
-.val-lbl { display:inline-flex; align-items:center; gap:6px; font-size:11px; color:var(--text-muted); text-transform:uppercase; letter-spacing:.05em; font-weight:600; }
-.val-select { font:inherit; font-size:12px; text-transform:none; letter-spacing:0; padding:3px 8px; border:1px solid var(--border); border-radius:6px; background:var(--bg-elevated); color:var(--text-primary); }
-.val-check { display:inline-flex; align-items:center; gap:6px; font-size:12px; color:var(--text-secondary); cursor:pointer; }
-.val-check input { margin:0; }
-.sch-validation-bar { display:flex; align-items:center; justify-content:space-between; gap:16px 20px; flex-wrap:wrap;
-  padding:12px 14px; margin-bottom:14px; border:1px solid var(--border); border-radius:10px; background:var(--bg-surface); }
-.sch-validation-bar .svb-label { display:flex; flex-direction:column; gap:2px; min-width:0; }
-.sch-validation-bar .svb-title { font-size:13px; font-weight:640; color:var(--text-primary); }
-.sch-validation-bar .svb-hint { font-size:11.5px; color:var(--text-muted); }
-.sch-md { display:grid; grid-template-columns:minmax(190px,250px) 1fr; gap:18px; align-items:start; margin-top:6px; }
-@media (max-width:760px) { .sch-md { grid-template-columns:1fr; } }
-.sch-master { display:flex; flex-direction:column; gap:3px; min-width:0; }
-/* The list of types scrolls inside itself; the add-row above and imports below stay pinned. */
-.sch-type-list { display:flex; flex-direction:column; gap:3px; min-height:0; overflow-y:auto; max-height:340px; }
-.sch-type-item { display:flex; align-items:center; gap:8px; width:100%; text-align:left; background:none;
-  border:1px solid transparent; border-radius:8px; padding:7px 9px; cursor:pointer; font:inherit; color:var(--text-primary); }
-.sch-type-item:hover { background:var(--bg-elevated); }
-.sch-type-item.sel { background:color-mix(in srgb,var(--accent) 12%,transparent); border-color:color-mix(in srgb,var(--accent) 34%,transparent); }
-.sch-type-item .nm { font-family:var(--font-mono); font-size:13px; color:var(--accent); flex:1; min-width:0;
-  overflow:hidden; text-overflow:ellipsis; white-space:nowrap; }
-.sch-type-badges { display:inline-flex; gap:3px; flex-wrap:wrap; justify-content:flex-end; }
-.sch-empty-list { color:var(--text-muted); font-size:12.5px; font-style:italic; padding:14px 6px; text-align:center; }
-.sch-detail { min-width:0; }
-.sch-detail-empty { color:var(--text-muted); font-size:13px; font-style:italic; padding:26px 20px; text-align:center;
-  border:1px dashed var(--border); border-radius:10px; }
-.sch-detail-head { display:flex; align-items:center; gap:10px; min-height:var(--sch-head-h);
-  margin-bottom:14px; padding-bottom:10px; border-bottom:1px solid var(--border); box-sizing:content-box; }
-.sch-detail-head .dt { font-family:var(--font-mono); font-size:15px; color:var(--accent); font-weight:600; flex:1; min-width:0;
-  overflow:hidden; text-overflow:ellipsis; }
-.sch-detail-head .acts { display:flex; gap:4px; flex-shrink:0; }
-/* Pinned above the list: a bottom rule, not a top one, because it now heads the column.
-
-   It and the detail pane's head are the two column headers, side by side, so they share one height and
-   one bottom margin — otherwise their rules sit at different y and the two columns read as misaligned
-   even though the grid starts them at the same top edge. --sch-head-h is that shared height; changing
-   it moves both. */
-.sch-md { --sch-head-h:34px; }
-.sch-add-row { display:flex; gap:6px; align-items:center; min-height:var(--sch-head-h);
-  margin-bottom:14px; padding-bottom:10px; border-bottom:1px solid var(--border); box-sizing:content-box; }
-.sch-add-row input { flex:1; min-width:0; }
-.sch-add-btn { display:grid; place-items:center; flex:none; width:30px; height:30px; padding:0;
-  border:1px solid var(--border); border-radius:8px; background:var(--bg-primary);
-  color:var(--accent); cursor:pointer; }
-.sch-add-btn:hover:not(:disabled) { border-color:var(--accent); }
-.sch-add-btn:disabled { color:var(--text-muted); cursor:not-allowed; opacity:.6; }
-.sch-add-btn:focus-visible { outline:2px solid var(--accent); outline-offset:2px; }
-/* Same row, but it heads the detail pane's foot rather than the list's head: rule on top, not bottom. */
-.sch-add-prop { margin-bottom:0; padding-bottom:0; border-bottom:none;
-  margin-top:10px; padding-top:10px; border-top:1px solid var(--border); }
-.sch-add-prop input { max-width:260px; }
-.sch-add-imports { display:flex; gap:6px; flex-wrap:wrap; margin-top:10px; padding-top:8px;
-  border-top:1px solid var(--border-muted); }
-.prop-caret { color:var(--text-muted); flex-shrink:0; display:inline-flex; }
-/* One coherent text scale for the tab: guidance, section labels, inline messages.
-   Every section label reads the same and every hint hangs off it the same way — the delimiter is an
-   em dash in all of them, where it used to be parentheses in some and a dash in others. */
-.sch-hint { font-size:11px; font-weight:400; text-transform:none; letter-spacing:0; color:var(--text-muted); }
-.sch-section-label { font-size:11px; font-weight:700; letter-spacing:.06em; text-transform:uppercase; color:var(--text-muted); margin-bottom:8px; }
-/* One rhythm between sections of the detail pane. They were spaced by whatever each block's own
-   margins happened to add up to, so the gaps above "Tag suggestions" and "Property schemas" differed
-   by several pixels for no reason a reader could infer. */
-.sch-detail .sch-section-label,
-.sch-detail > .field > label { margin-top:16px; }
-.sch-detail > .field:first-of-type > label,
-.sch-detail .sch-section-label:first-of-type { margin-top:0; }
-/* The two retention windows sit side by side and wrap only on a genuinely narrow pane.
-   .field must be given a BASIS: it is a flex column, so its intrinsic width is its widest child, and the
-   chrono hint under the second input is a long sentence — left to size itself that field claimed the whole
-   row and both stacked. Verified by measurement, not by looking at the CSS: the first attempt reported
-   two inputs with the labels and placeholders all correct, and they were one above the other. */
-.ret-row { display:flex; gap:14px; flex-wrap:wrap; align-items:flex-start; }
-.ret-row .field { flex:1 1 190px; min-width:0; max-width:260px; }
-.ret-row input { max-width:150px; }
-.sch-msg { font-size:12px; margin-top:6px; }
-.sch-msg.err { color:var(--error); }
-.sch-msg.ok  { color:var(--success); }
-.sch-type-badges .badge { font-size:9px; }
-`;
+import { SCHEMA_MD_STYLES } from './schema-styles';
+import { SchemaTypeEditorComponent } from './schema-type-editor.component';
 
 @Component({
   selector: 'app-space-schema-tab',
   standalone: true,
-  imports: [CommonModule, FormsModule, TranslocoPipe, PhIconComponent, ModalDirective, HscrollTopDirective, ErrorStateComponent],
+  imports: [SchemaTypeEditorComponent, CommonModule, FormsModule, TranslocoPipe, PhIconComponent, ModalDirective, ErrorStateComponent],
   changeDetection: ChangeDetectionStrategy.OnPush,
   styles: [SPACE_DIALOG_STYLES, SCHEMA_MD_STYLES],
   template: `
@@ -281,230 +196,16 @@ const SCHEMA_MD_STYLES = `
           </span>
         </div>
 
-        @if (state.typeLibRef(kt,name); as libRef) {
-          <!-- Linked library schema — editable only after unlinking; shown read-only meanwhile. -->
-          <div style="display:flex;align-items:center;gap:10px;padding:4px 0;color:var(--text-secondary);font-size:13px;">
-            <ph-icon name="bookmarks" [size]="16" style="color:var(--accent);flex-shrink:0;"/>
-            <span>{{ 'spaces.schema.libRef.linkedHint' | transloco: {name: libRef} }}</span>
-            <button class="btn btn-secondary btn-sm" type="button" style="margin-left:auto;flex-shrink:0;"
-              (click)="unlinkType(kt,name)" [attr.title]="'spaces.schema.libRef.unlinkTitle' | transloco">{{ 'spaces.schema.libRef.unlinkButton' | transloco }}</button>
-          </div>
-          <!-- Read-only view of the linked entry's properties, so you can see what the type enforces
-               without unlinking first. -->
-          @if (linkedProps(kt,name); as props) {
-            @if (props.length) {
-              <div class="sch-section-label" style="margin-top:12px;">{{ 'spaces.schema.propertySchemas' | transloco }}</div>
-              <div class="tablewrap">
-                <table>
-                  <thead>
-                    <tr>
-                      <th style="width:150px;">{{ 'spaces.schema.propTable.property' | transloco }}</th>
-                      <th style="width:80px;">{{ 'spaces.schema.propTable.type' | transloco }}</th>
-                      <th>{{ 'spaces.schema.propTable.constraints' | transloco }}</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    @for (p of props; track p.key) {
-                      <tr>
-                        <td style="font-family:var(--font-mono);font-size:12.5px;">{{ p.key }}</td>
-                        <td>{{ p.s.type || '—' }}</td>
-                        <td style="color:var(--text-secondary);font-size:12.5px;">{{ propConstraintSummary(p.s) }}</td>
-                      </tr>
-                    }
-                  </tbody>
-                </table>
-              </div>
-            } @else {
-              <div class="sch-hint" style="margin-top:8px;">{{ 'spaces.schema.libRef.noProps' | transloco }}</div>
-            }
-          }
-        } @else {
-          <!-- Naming pattern (entity only) -->
-          @if (kt === 'entity') {
-            <div class="field" style="margin:0 0 12px;">
-              <label>{{ 'spaces.schema.namingPattern' | transloco }} <span class="sch-hint">{{ 'spaces.schema.namingPatternHint' | transloco }}</span></label>
-              <input type="text" [(ngModel)]="state.typeState(kt,name).namingPattern" [placeholder]="'spaces.schema.namingPatternPlaceholder' | transloco" style="max-width:320px;" />
-            </div>
-          }
-          <!-- Retention — the SCHEMA tier of record > schema > space, and the control the Danger Zone, the
-               integration guide and the API have all been pointing at. It belongs here, beside the type's
-               other rules, rather than in a second parallel map an operator has to know exists.
-
-               The hint names what an empty field inherits, with the space default's actual number in it: the
-               operator who asked for this said the old arrangement was a convention they had to know, and
-               "inherit" without saying inherit-WHAT is the same failure one level down.
-
-               NOTE: no backticks anywhere in this comment — one kills the whole template string and the
-               error then points at @Component. -->
-          <div class="sch-section-label">{{ 'spaces.schema.retention.label' | transloco }}
-            <!-- The inherited number is THIS collection's bucket, not one space-wide figure: the space tier is
-                 five windows, and naming the wrong one would be worse than naming none. -->
-            <span class="sch-hint">
-              @if (spaceWindow(kt); as days) {
-                {{ 'spaces.schema.retention.hintSpace' | transloco: { days } }}
-              } @else {
-                {{ 'spaces.schema.retention.hintNoSpace' | transloco }}
-              }
-            </span>
-          </div>
-          <div class="ret-row">
-            <div class="field" style="margin:0;">
-              <label>{{ 'spaces.schema.retention.days' | transloco }}</label>
-              <input type="number" min="1" step="1" [(ngModel)]="state.typeState(kt,name).retentionDays"
-                [placeholder]="'spaces.schema.retention.inherit' | transloco" />
-            </div>
-            <!-- chrono only, because that is the only collection whose sweep implements it. Offering it on
-                 the others would store a number that never fires. -->
-            @if (kt === 'chrono') {
-              <div class="field" style="margin:0;">
-                <label>{{ 'spaces.schema.retention.contentDays' | transloco }}</label>
-                <input type="number" min="1" step="1" [(ngModel)]="state.typeState(kt,name).retentionContentDays"
-                  [placeholder]="'spaces.schema.retention.never' | transloco" />
-                <div class="sch-hint" style="margin-top:3px;">{{ 'spaces.schema.retention.contentDaysHint' | transloco }}</div>
-              </div>
-            }
-          </div>
-          <!-- The server CLAMPS a content window that is not strictly inside the delete window (it would
-               never fire), so without this the field would accept a number and silently do nothing. -->
-          @if (contentWindowNeverFires(kt,name); as total) {
-            <div class="sch-msg err">{{ 'spaces.schema.retention.contentTooLate' | transloco: { total } }}</div>
-          }
-          <!-- Per-type tag suggestions were retired here. The editor reached nothing: not the Brain
-               record forms (they suggest from tags already in use) and not the schema guidance sent to
-               MCP clients. Offering a control that does nothing is the dishonesty the Models rebuild
-               spent four PRs removing, and it is the same reasoning that retired the space-wide list
-               in #365. Stored values are preserved — see the note on TypeSchema.tagSuggestions. -->
-          <!-- Property schemas -->
-          <!-- Every other section in this pane explains itself; this one did not, and it is the one
-               doing the most work. The hint also points at the control that decides enforcement,
-               which is a whole panel away at the top of the tab. -->
-          <div class="sch-section-label">{{ 'spaces.schema.propertySchemas' | transloco }}
-            <span class="sch-hint">{{ 'spaces.schema.propertySchemasHint' | transloco }}</span></div>
-          <div class="table-wrapper" hscrollTop style="margin-bottom:0;">
-            <table class="prop-table" style="margin-bottom:0;">
-              <thead>
-                <tr>
-                  <th style="width:30px;"></th>
-                  <th style="width:150px;">{{ 'spaces.schema.propTable.property' | transloco }}</th>
-                  <th style="width:80px;">{{ 'spaces.schema.propTable.type' | transloco }}</th>
-                  <th>{{ 'spaces.schema.propTable.constraints' | transloco }}</th>
-                  <th style="width:40px;"></th>
-                </tr>
-              </thead>
-              <tbody>
-                @for (p of state.typeState(kt,name).propertySchemas; track p.key) {
-                  <tr class="prop-row" [class.prow-open]="state.isPropExpanded(kt,name,p.key)"
-                    (click)="state.togglePropExpand(kt,name,p.key)">
-                    <td><span class="prop-caret"><ph-icon [name]="state.isPropExpanded(kt,name,p.key) ? 'caret-up' : 'caret-down'" [size]="13"/></span></td>
-                    <td>
-                      <div style="display:flex;align-items:center;gap:7px;">
-                        <span style="font-family:var(--font-mono);font-size:12px;">{{ p.key }}</span>
-                        <label class="req-toggle" [class.is-req]="p.s.required" (click)="$event.stopPropagation()">
-                          <input type="checkbox" [checked]="p.s.required" (change)="p.s.required = !p.s.required" style="pointer-events:none;" />
-                          {{ 'spaces.schema.propDetail.required' | transloco }}
-                        </label>
-                      </div>
-                    </td>
-                    <td><span class="badge badge-gray">{{ p.s.type ?? 'any' }}</span></td>
-                    <td style="font-size:11px;color:var(--text-muted);">
-                      @if (p.s.enum?.length) { <span class="badge badge-gray" style="margin-right:3px">enum {{ p.s.enum!.length }}</span> }
-                      @if (p.s.minimum!==undefined) { <span style="margin-right:4px;">min:{{ p.s.minimum }}</span> }
-                      @if (p.s.maximum!==undefined) { <span style="margin-right:4px;">max:{{ p.s.maximum }}</span> }
-                      @if (p.s.pattern) { <span style="margin-right:4px;">pattern</span> }
-                      @if (p.s.default!==undefined) { <span style="margin-right:4px;">default:{{ p.s.default }}</span> }
-                      @if (p.s.mergeFn) { <span class="badge badge-blue">{{ p.s.mergeFn }}</span> }
-                    </td>
-                    <td (click)="$event.stopPropagation()">
-                      <button class="icon-btn danger" type="button" (click)="state.removeProp(kt,name,p.key)" [attr.title]="'common.remove' | transloco"><ph-icon name="x" [size]="14"/></button>
-                    </td>
-                  </tr>
-                  @if (state.isPropExpanded(kt,name,p.key)) {
-                    <tr class="prop-expand-row" (click)="$event.stopPropagation()">
-                      <td colspan="5" style="padding:0;">
-                        <div class="pdet">
-                          <div class="pdet-fields">
-                            <div class="field" style="margin:0;">
-                              <label>{{ 'spaces.schema.propDetail.type' | transloco }}</label>
-                              <select [(ngModel)]="p.s.type">
-                                <option [ngValue]="undefined">any</option>
-                                <option value="string">string</option>
-                                <option value="number">number</option>
-                                <option value="boolean">boolean</option>
-                                <option value="date">date</option>
-                              </select>
-                            </div>
-                            <div class="field" style="margin:0;">
-                              <label>{{ 'spaces.schema.propDetail.default' | transloco }}</label>
-                              <input type="text" [(ngModel)]="p.s.default" placeholder="—" />
-                            </div>
-                            <div class="field" style="margin:0;">
-                              <label>{{ 'spaces.schema.propDetail.mergeFn' | transloco }}</label>
-                              <select [(ngModel)]="p.s.mergeFn">
-                                <option [ngValue]="undefined">—</option>
-                                <option value="avg">avg</option><option value="min">min</option>
-                                <option value="max">max</option><option value="sum">sum</option>
-                                <option value="and">and</option><option value="or">or</option>
-                                <option value="xor">xor</option>
-                              </select>
-                            </div>
-                            @if (p.s.type==='string'||p.s.type===undefined) {
-                              <div class="field" style="margin:0;">
-                                <label>{{ 'spaces.schema.propDetail.pattern' | transloco }} <span class="sch-hint">{{ 'spaces.schema.propDetail.patternHint' | transloco }}</span></label>
-                                <input type="text" [(ngModel)]="p.s.pattern" placeholder="^[A-Z].*" />
-                              </div>
-                            }
-                            @if (p.s.type==='number'||p.s.type===undefined) {
-                              <div class="field" style="margin:0;">
-                                <label>{{ 'spaces.schema.propDetail.min' | transloco }}</label>
-                                <input type="number" [(ngModel)]="p.s.minimum" placeholder="—" />
-                              </div>
-                              <div class="field" style="margin:0;">
-                                <label>{{ 'spaces.schema.propDetail.max' | transloco }}</label>
-                                <input type="number" [(ngModel)]="p.s.maximum" placeholder="—" />
-                              </div>
-                            }
-                          </div>
-                          @if (p.s.type !== 'boolean') {
-                            <div class="pdet-full">
-                              <div class="field" style="margin:0;">
-                                <label>{{ 'spaces.schema.propDetail.enumValues' | transloco }} <span class="sch-hint">{{ 'spaces.schema.propDetail.enumHint' | transloco }}</span></label>
-                                <div class="chip-wrap">
-                                  @for (ev of (p.s.enum??[]); track ev) {
-                                    <span class="chip">{{ ev }}<button type="button" class="chip-rm" (click)="state.removeEnumVal(kt,name,p.key,ev)"><ph-icon name="x" [size]="12"/></button></span>
-                                  }
-                                  <input type="text" class="chip-field" [(ngModel)]="p._enumInput"
-                                    [placeholder]="'spaces.schema.propDetail.enumPlaceholder' | transloco" (keydown)="state.onEnumKey($event,kt,name,p.key)" />
-                                </div>
-                              </div>
-                            </div>
-                          }
-                        </div>
-                      </td>
-                    </tr>
-                  }
-                } @empty {
-                  <tr>
-                    <td colspan="5" style="padding:24px 0;text-align:center;color:var(--text-muted);font-size:13px;font-style:italic;">
-                      {{ 'spaces.schema.noProps' | transloco }}
-                    </td>
-                  </tr>
-                }
-              </tbody>
-            </table>
-          </div>
-          <!-- add property — the same inline [input][+] affordance as the add-type row, so the two
-               "add something" controls on this tab read and behave identically. -->
-          <div class="sch-add-row sch-add-prop">
-            <input type="text" [(ngModel)]="state.typeState(kt,name)._newPropInput" [placeholder]="'spaces.schema.newPropNamePlaceholder' | transloco"
-              [attr.aria-label]="'spaces.schema.addPropertyButton' | transloco"
-              (keydown.enter)="$event.preventDefault();state.addProp(kt,name)" />
-            <button class="sch-add-btn" type="button"
-              (click)="state.addProp(kt,name)" [disabled]="!state.typeState(kt,name)._newPropInput.trim()"
-              [attr.title]="'spaces.schema.addPropertyButton' | transloco" [attr.aria-label]="'spaces.schema.addPropertyButton' | transloco">
-              <ph-icon name="plus-circle" [size]="18"/>
-            </button>
-          </div>
-        }
+        <!-- The editor body is a shared component: the Brain Overview opens the SAME one in a dialog, so a
+             one-field schema change no longer means a trip to Space Settings and back. The header above
+             keeps the library and delete actions deliberately — see the component for why. -->
+        <app-schema-type-editor
+          [knowledgeType]="kt"
+          [draft]="state.typeState(kt,name)"
+          [libRef]="state.typeLibRef(kt,name)"
+          [linkedProps]="linkedProps(kt,name)"
+          [spaceWindowDays]="spaceWindow(kt)"
+          (unlink)="unlinkType(kt,name)" />
       } @else {
         <div class="sch-detail-empty">{{ kt === 'edge' ? ('spaces.schema.detail.emptyEdge' | transloco) : ('spaces.schema.detail.empty' | transloco) }}</div>
       }
